cleanup, inline docs, cascading types
